backend/app/ml/requirements_training.txt
# Training Dependencies
# Install with: pip install -r requirements_training.txt

tensorflow>=2.16.0
keras>=3.10.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.4.0
pillow>=10.4.0
pandas>=2.2.0

# Optional but recommended
opencv-python>=4.9.0
scikit-image>=0.23.0
